requirements.txt
flask
bootstrap-flask
pymysql
python-dotenv